brought out OffersByLocationSection in separate file
