+++ src/components/molecules/index.ts
@@ -17,6 +17,7 @@ import ReviewForm from './review-form/review-form.tsx';
 import RatingButton from './rating-button/rating-button.tsx';
 import StarInput from './star-input/star-input.tsx';
 import Map from './map/map.tsx';
+import OffersByLocationSection from './offers-by-location-section/offers-by-location-section.tsx';
 
 export {
   CityTabs,
@@ -38,4 +39,5 @@ export {
   RatingButton,
   StarInput,
   Map,
+  OffersByLocationSection,
 };
